Local ANS Decoder Complete
